[FEAT] BannerCarousel 분리 | Separate BannerCarousel  TALEARNT-416
